Employee-Service/requirements.txt
# FastAPI Employee Service Requirements
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.32.1
starlette==0.44.0

# Database & ORM
sqlalchemy==2.0.35
alembic==1.16.4
asyncpg==0.30.0

# Security & Authentication
PyJWT==2.10.1

# Data Validation & Settings
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0

# Configuration & Environment
python-decouple==3.8
python-dotenv==1.1.1

# CORS & Middleware
fastapi-cors==0.0.6

# HTTP Client for Auth Service integration
httpx==0.28.1

# Event Bus (optional - NATS or Kafka)
nats-py==2.8.0

# Observability
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-auto-instrumentation==0.41b0

# Logging
colorama==0.4.6

# Development & Testing
pytest==8.4.1
pytest-asyncio==1.1.0

# Additional Utilities
typing-extensions==4.12.2
